requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --extra=dev --output-file=requirements-dev.txt pyproject.toml
#
beautifulsoup4==4.14.3
    # via
    #   altbrow (pyproject.toml)
    #   mf2py
build==1.4.3
    # via altbrow (pyproject.toml)
certifi==2026.2.25
    # via requests
charset-normalizer==3.4.7
    # via requests
click==8.3.2
    # via mkdocs
colorama==0.4.6
    # via
    #   build
    #   click
    #   mkdocs
    #   pytest
dnspython==2.8.0
    # via altbrow (pyproject.toml)
extruct==0.18.0
    # via altbrow (pyproject.toml)
filelock==3.25.2
    # via tldextract
ghp-import==2.1.0
    # via mkdocs
griffelib==2.0.2
    # via mkdocstrings-python
html-text==0.7.1
    # via extruct
html5lib==1.1
    # via
    #   mf2py
    #   pyrdfa3
idna==3.11
    # via
    #   requests
    #   tldextract
iniconfig==2.3.0
    # via pytest
jinja2==3.1.6
    # via
    #   mkdocs
    #   mkdocstrings
jstyleson==0.0.2
    # via extruct
lxml==6.0.4
    # via
    #   altbrow (pyproject.toml)
    #   extruct
    #   html-text
    #   lxml-html-clean
lxml-html-clean==0.4.4
    # via
    #   extruct
    #   html-text
markdown==3.10.2
    # via
    #   mkdocs
    #   mkdocs-autorefs
    #   mkdocstrings
    #   pymdown-extensions
markupsafe==3.0.3
    # via
    #   jinja2
    #   mkdocs
    #   mkdocs-autorefs
    #   mkdocstrings
maxminddb==3.1.1
    # via altbrow (pyproject.toml)
mergedeep==1.3.4
    # via
    #   mkdocs
    #   mkdocs-get-deps
mf2py==2.0.1
    # via extruct
mkdocs==1.6.1
    # via
    #   altbrow (pyproject.toml)
    #   mkdocs-autorefs
    #   mkdocstrings
mkdocs-autorefs==1.4.4
    # via
    #   mkdocstrings
    #   mkdocstrings-python
mkdocs-get-deps==0.2.2
    # via mkdocs
mkdocstrings[python]==1.0.3
    # via
    #   altbrow (pyproject.toml)
    #   mkdocstrings-python
mkdocstrings-python==2.0.3
    # via mkdocstrings
packaging==26.0
    # via
    #   build
    #   mkdocs
    #   pytest
pathspec==1.0.4
    # via mkdocs
platformdirs==4.9.6
    # via mkdocs-get-deps
pluggy==1.6.0
    # via pytest
pygments==2.20.0
    # via pytest
pymdown-extensions==10.21.2
    # via mkdocstrings
pyparsing==3.3.2
    # via rdflib
pyproject-hooks==1.2.0
    # via build
pyrdfa3==3.6.5
    # via extruct
pytest==9.0.3
    # via altbrow (pyproject.toml)
python-dateutil==2.9.0.post0
    # via ghp-import
pyyaml==6.0.3
    # via
    #   altbrow (pyproject.toml)
    #   mkdocs
    #   mkdocs-get-deps
    #   pymdown-extensions
    #   pyyaml-env-tag
pyyaml-env-tag==1.1
    # via mkdocs
rdflib==7.6.0
    # via
    #   altbrow (pyproject.toml)
    #   extruct
    #   pyrdfa3
requests==2.33.1
    # via
    #   altbrow (pyproject.toml)
    #   mf2py
    #   pyrdfa3
    #   requests-file
    #   tldextract
requests-file==3.0.1
    # via tldextract
ruff==0.15.10
    # via altbrow (pyproject.toml)
six==1.17.0
    # via
    #   html5lib
    #   python-dateutil
soupsieve==2.8.3
    # via beautifulsoup4
tldextract==5.3.1
    # via altbrow (pyproject.toml)
typing-extensions==4.15.0
    # via beautifulsoup4
urllib3==2.6.3
    # via requests
w3lib==2.4.1
    # via extruct
watchdog==6.0.0
    # via mkdocs
webencodings==0.5.1
    # via html5lib
